move tests for logging of master to separate file
